fixes
- u16 -> s8 trainermon struct
- switch as an option again
- quick claw for dwebble
- brock and misty mon count to 4 and 5, resp.

+++ include/battle.h
@@ -260,7 +260,7 @@ struct TrainerMonNoItemCustomMoves
 
 struct TrainerMonItemCustomMoves
 {
-    u16 lvl;
+    s8 lvl;
     u16 species;
 	u8 	ability;
 	u8 	nature;
